+++ reactive-grpc-common/src/main/java/com/salesforce/reactivegrpccommon/ReactiveGrpcGenerator.java
@@ -23,7 +23,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
- * A protoc generator for generating ReactiveX 2.0 bindings for gRPC.
+ * Abstract class for protoc generators generating Reactive Streams bindings for gRPC.
  */
 public abstract class ReactiveGrpcGenerator extends Generator {
