+++ reactive-grpc-common/src/main/java/com/salesforce/reactivegrpccommon/testing/Sequence.java
@@ -5,7 +5,7 @@
  *  For full license text, see LICENSE.txt file in the repo root  or https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.salesforce.reactivegrpccommon;
+package com.salesforce.reactivegrpccommon.testing;
 
 import javax.annotation.Nonnull;
 import java.util.Iterator;
